chore: Update npm dependencies to latest versions
This commit updates the npm dependencies in the package.json file to their latest stable versions. Specifically, the "hono" dependency is updated from version 4.5.5 to 4.5.6, and the "@cloudflare/workers-types" dependency is updated from version 4.20240806.0 to 4.20240815.0. This ensures that the project is using the most up-to-date dependencies and takes advantage of any bug fixes or improvements provided by the latest versions.