```text chore: Add build script to package.json
This commit adds a new "build" script to the package.json file. The build script uses the "bun" command to build the source code located in src/index.ts and outputs the built files to the dist directory. This script can be used to generate a production-ready build of the project.